Detallitos en noticias y actualizar datos

+++ app/Http/Controllers/MantenimientoNoticiaController.php
@@ -9,6 +9,7 @@ use Redirect;
 use Yajra\Datatables\Datatables;
 use Laracasts\Flash\Flash;
 use avaa\Http\Requests\NoticiaRequest;
+use avaa\Http\Requests\NoticiaActualizarRequest;
 use Validator;
 use Illuminate\Validation\Rule;
 
@@ -42,13 +43,14 @@ class MantenimientoNoticiaController extends Controller
         $noticia->url_imagen = '/images/noticias/'.$name;
         $noticia->al_carrousel = ($request->destacada=='1') ? 1 : 0;
         $tipo= ( 'noticia' === $noticia->tipo ) ? 'Noticia' : 'Miembro Institucional';
+
         if($noticia->save())
         {
             flash('La publicación tipo: '.$tipo.' fue registrada exitosamente.','success')->important();
         }
         else
         {
-            flash('Ha ocurrido un error al registrar el publicación')->error()->important();
+            flash('Disculpe, ha ocurrido un error al registrar la publicación.')->error()->important();
         }
         return redirect()->route('noticia.index');
         
@@ -64,19 +66,20 @@ class MantenimientoNoticiaController extends Controller
         $noticia = Noticia::find($id);
         if(is_null($noticia))
         {
-            flash('El Archivo solicitado no ha sido encontrado')->error()->important();
+            flash('Disculpe, el archivo solicitado no ha sido encontrado.')->error()->important();
             return back();
         }
         return view('sisbeca.noticias.editarNoticia')->with('noticia',$noticia);
     }
 
-    public function update(NoticiaRequest $request, $id)
+    public function update(NoticiaActualizarRequest $request, $id)
     {
         $noticia = Noticia::find($id);
         $file= $request->file('url_imagen');
         if($request->tipo==='miembroins')
         {
-            $errores = Validator::make($request->all(), [
+            $errores = Validator::make($request->all(), 
+                [
                 'url_articulo' => 'required|url',
                 'email_contacto' => [
                     Rule::unique('noticias')->ignore($noticia->id),
@@ -86,27 +89,35 @@ class MantenimientoNoticiaController extends Controller
             ])->validate();
         }
         
-        $es_archivo=is_file(public_path() .$noticia->url_imagen);
-        if($es_archivo)
+        if($request->url_imagen)
         {
-            unlink(public_path() .$noticia->url_imagen);
+            $es_archivo=is_file(public_path() .$noticia->url_imagen);
+            if($es_archivo)
+            {
+                unlink(public_path() .$noticia->url_imagen);
+            }
+            $name = 'noticiasAVAA_' . time() . '.' . $file->getClientOriginalExtension();
+            $path = public_path() . '/images/noticias/';
+            $file->move($path, $name);
         }
-        $name = 'noticiasAVAA_' . time() . '.' . $file->getClientOriginalExtension();
-        $path = public_path() . '/images/noticias/';
-        $file->move($path, $name);
+        
 
         $noticia->fill($request->all());
         $noticia->slug = Noticia::getSlug($noticia->titulo);
         $noticia->user_id = \Auth::user()->id;
-        $noticia->url_imagen = '/images/noticias/'.$name;
+        if($request->url_imagen)
+        {
+            $noticia->url_imagen = '/images/noticias/'.$name;
+        }
         $noticia->al_carrousel = ($request->destacada=='1') ? 1 : 0;
+
         if($noticia->save())
         {
             flash('La publicación fue actualizada exitosamente.','success')->important();
         }
         else
         {
-            flash('Ha ocurrido un error al actualizar publicación.')->error()->important();
+            flash('Disculpe, ha ocurrido un error al actualizar la publicación.')->error()->important();
         }
         return  redirect()->route('noticia.index');
     }
@@ -116,7 +127,7 @@ class MantenimientoNoticiaController extends Controller
         $noticia= Noticia::find($id);
         if(is_null($noticia))
         {
-            flash('El Archivo solicitado no ha sido encontrado.')->error()->important();
+            flash('Disculpe, el archivo solicitado no ha sido encontrado.')->error()->important();
             return back();
         }
         if($noticia->delete())
@@ -124,11 +135,11 @@ class MantenimientoNoticiaController extends Controller
             $es_archivo=file_exists(public_path() .$noticia->url_imagen);
             if($es_archivo)
                      unlink(public_path() .$noticia->url_imagen);
-            flash('La publicación ha sido eliminada exitosamente.','success')->important();
+            flash('La publicación fue eliminada exitosamente.','success')->important();
         }
         else
         {
-            flash('Ha Ocurrido un error al eliminar publicación.')->error()->important();
+            flash('Disculpe, ha ocurrido un error al eliminar publicación.')->error()->important();
         }
         return  redirect()->route('noticia.index');
     }
